invites: splitting from roller, removing unused code, adding status
