+++ src/components/L2/Headers/L2PointHeader.tsx
@@ -10,24 +10,24 @@ import HeaderButton from './HeaderButton';
 import AccountsDropdown from '../Dropdowns/AccountsDropdown';
 import './L2PointHeader.scss';
 import LayerIndicator from '../LayerIndicator';
+import { useInviteStore } from 'views/Invite/useInvites';
 
 export interface L2PointHeaderProps {
-  numInvites?: number;
   hideHome?: boolean;
-  hideTimer?: boolean;
   hideInvites?: boolean;
   showMigrate?: boolean;
 }
 
 const L2PointHeader = ({
-  numInvites = 0,
   hideHome = false,
-  hideTimer = false,
   hideInvites = false,
   showMigrate = false,
 }: L2PointHeaderProps) => {
-  const { point, pointList, invitesLoading } = useRollerStore();
+  const { point, pointList } = useRollerStore();
+  const { invites, loading: invitesLoading } = useInviteStore();
   const { pointCursor, setPointCursor }: any = usePointCursor();
+  const numInvites = invites[point.value]?.length;
+  debugger;
 
   const { popTo, push, names }: any = useHistory();
