+++ src/lib/useRoller.ts
@@ -1,9 +1,8 @@
-import { useCallback, useEffect, useMemo, useRef, useState } from 'react';
+import { useCallback, useEffect, useMemo, useState } from 'react';
 import * as azimuth from 'azimuth-js';
 import * as need from 'lib/need';
 import SecureLS from 'secure-ls';
 
-import { Invite } from 'lib/types/Invite';
 import {
   deriveNetworkSeedFromUrbitWallet,
   attemptNetworkSeedDerivation,
@@ -13,24 +12,17 @@ import { useNetwork } from 'store/network';
 import { usePointCursor } from 'store/pointCursor';
 import { EMPTY_POINT, useRollerStore } from 'store/rollerStore';
 import { useWallet } from 'store/wallet';
-import {
-  getStoredInvites,
-  setStoredInvite,
-  setStoredInvites,
-} from 'store/storage/roller';
 import { usePointCache } from 'store/pointCache';
 
 import { getUpdatedPointMessage, isPlanet, toL1Details } from './utils/point';
 import { convertToInt } from './convertToInt';
 import { isDevelopment, isMainnet, isRopsten } from './flags';
 import {
-  generateInviteWallet,
-  getPendingSpawns,
   submitL2Transaction,
   registerProxyAddress,
   isL2Spawn,
 } from './utils/roller';
-import { POINT_DOMINIONS, ROLLER_HOSTS, TEN_SECONDS } from './constants';
+import { ROLLER_HOSTS, TEN_SECONDS } from './constants';
 
 import {
   Config,
@@ -39,7 +31,6 @@ import {
   RollerRPCAPI,
   Options,
   EthAddress,
-  UnspawnedPoints,
 } from '@urbit/roller-api';
 
 import Point, { PointField, Points } from './types/Point';
@@ -60,18 +51,6 @@ interface UpdateParams {
   intervalTime?: number;
 }
 
-const inviteTemplate = (
-  planet: number,
-  ticket = '',
-  owner = '',
-  hash = ''
-): Invite => ({
-  planet,
-  ticket,
-  owner,
-  hash,
-});
-
 export default function useRoller() {
   const {
     wallet,
@@ -89,22 +68,16 @@ export default function useRoller() {
   const {
     point,
     points,
-    pointList,
     nextQuotaTime,
-    addInvite,
     setNextBatchTime,
     setNextQuotaTime,
     setPendingTransactions,
-    setInvites,
-    setInviteGeneratingNum,
-    setInvitesLoading,
     setModalText,
     setPoints,
     updatePoint,
     storePendingL1Txn,
     deletePendingL1Txn,
   } = useRollerStore();
-  const generatingInvites = useRef(false);
   const [config, setConfig] = useState<Config | null>(null);
 
   const options: Options = useMemo(() => {
@@ -341,247 +314,6 @@ export default function useRoller() {
       walletType,
       web3,
       quotaReached,
-    ]
-  );
-
-  const generateInviteInfo = async (planet: number, _authToken: string) => {
-    const { ticket, inviteWallet } = await generateInviteWallet(
-      planet,
-      _authToken
-    );
-
-    return inviteTemplate(planet, ticket, inviteWallet.ownership.keys.address);
-  };
-
-  const getInvites = useCallback(async () => {
-    if (generatingInvites.current) {
-      return;
-    }
-
-    generatingInvites.current = true;
-    try {
-      setInvitesLoading(true);
-      const curPoint: number = Number(need.point(pointCursor));
-      const _authToken = authToken.getOrElse(null);
-      const _contracts = contracts.getOrElse(null);
-
-      if (_authToken && _contracts) {
-        const spawnedPoints = point.isL2Spawn
-          ? await api.getSpawned(curPoint)
-          : [];
-
-        const newPending = await api.getPendingByShip(curPoint);
-        setPendingTransactions(newPending);
-        const pendingSpawns = getPendingSpawns(newPending);
-
-        const availablePoints = await azimuth.azimuth.getUnspawnedChildren(
-          _contracts,
-          curPoint
-        );
-
-        const outgoingPoints = pointList
-          .filter(
-            ({ value, isPlanet }) => isPlanet && availablePoints.includes(value)
-          )
-          .map(({ value }) => value);
-
-        const invitePlanets = spawnedPoints
-          .concat(
-            outgoingPoints.filter((p: number) => !spawnedPoints.includes(p))
-          )
-          .filter((p: number) => !pendingSpawns.has(p))
-          .sort();
-
-        setInvites(curPoint, []);
-
-        const newInvites: Invite[] = [];
-        // Iterate over all of the stored invites, generating wallet info as necessary
-        const storedInvites = getStoredInvites(ls);
-
-        for (let i = 0; i < invitePlanets.length; i++) {
-          setInviteGeneratingNum(i + 1);
-          const planet = invitePlanets[i];
-          const storedInvite = storedInvites[planet];
-          const invite =
-            storedInvite || (await generateInviteInfo(planet, _authToken));
-
-          // TODO: check if the invite point's owner still matches the deterministic wallet address
-          const planetInfo = await api.getPoint(planet);
-          setStoredInvite(ls, invite);
-
-          if (
-            invite.owner.toLowerCase() ===
-              planetInfo.ownership?.owner?.address &&
-            planetInfo.dominion === POINT_DOMINIONS.L2
-          ) {
-            addInvite(curPoint, invite);
-            newInvites.push(invite);
-          }
-
-          if (storedInvite) {
-            setStoredInvite(ls, invite);
-          }
-        }
-
-        setInvitesLoading(false);
-        return newInvites;
-      }
-    } catch (error) {
-      console.warn('ERROR GETTING INVITES', error);
-    }
-
-    generatingInvites.current = false;
-  }, [
-    api,
-    authToken,
-    contracts,
-    pointCursor,
-    ls,
-    point,
-    pointList,
-    addInvite,
-    setInvites,
-    setInvitesLoading,
-    setInviteGeneratingNum,
-    setPendingTransactions,
-  ]);
-
-  const generateInviteCodes = useCallback(
-    async (numInvites: number) => {
-      const _point = need.point(pointCursor);
-      const _contracts = contracts.getOrElse(null);
-      const _web3 = web3.getOrElse(null);
-      const _wallet = wallet.getOrElse(null);
-      const _authToken = authToken.getOrElse(null);
-      const _details = getDetails(_point);
-
-      if (!_contracts || !_web3 || !_wallet || !_authToken || !_details) {
-        // not using need because we want a custom error
-        throw new Error('Internal Error: Missing Contracts/Web3/Wallet');
-      }
-
-      // TODO: remove when UI disables management proxy from spawning invites
-      // if (Just.hasInstance(authMnemonic)) {
-      //   throw new Error(
-      //     "Auth key Error: A management mnemonic can't create invite codes"
-      //   );
-      // }
-      let planets: UnspawnedPoints = await api.getUnspawned(_point);
-
-      const pendingTxs = await api.getPendingByShip(_point);
-
-      const pendingSpawns = getPendingSpawns(pendingTxs);
-
-      planets = planets.filter((point: number) => !pendingSpawns.has(point));
-
-      const invites: Invite[] = [];
-
-      const proxy = point.getSpawnProxy();
-
-      if (proxy === undefined)
-        throw new Error("Error: Address doesn't match proxy");
-
-      const nonce = await api.getNonce({
-        ship: _point,
-        proxy,
-      });
-
-      for (let i = 0; i < numInvites && planets[i]; i++) {
-        setInviteGeneratingNum(i + 1);
-        const planet = planets[i];
-        const nonceInc = i + nonce;
-
-        const { ticket, inviteWallet } = await generateInviteWallet(
-          planet,
-          _authToken
-        );
-
-        await submitL2Transaction({
-          api,
-          address: _wallet.address,
-          wallet: _wallet,
-          ship: _point,
-          proxy,
-          nonce: nonceInc,
-          pointToSpawn: planet,
-          type: 'spawn',
-          walletType,
-          web3: _web3,
-          connector,
-        });
-
-        const networkSeed = await deriveNetworkSeedFromUrbitWallet(
-          inviteWallet,
-          1
-        );
-        await submitL2Transaction({
-          api,
-          wallet: _wallet,
-          ship: planet,
-          proxy: 'own',
-          nonce: 0,
-          networkSeed,
-          type: 'configureKeys',
-          walletType,
-          web3: _web3,
-          connector,
-        });
-
-        await registerProxyAddress(
-          api,
-          _wallet,
-          planet,
-          'own',
-          'manage',
-          1,
-          inviteWallet.management.keys.address,
-          walletType,
-          _web3,
-          connector
-        );
-
-        await submitL2Transaction({
-          api,
-          wallet: _wallet,
-          ship: planet,
-          proxy: 'own',
-          type: 'transferPoint',
-          nonce: 2,
-          address: inviteWallet.ownership.keys.address,
-          walletType,
-          web3: _web3,
-          connector,
-        });
-
-        invites.push({
-          hash: '',
-          ticket,
-          planet,
-          owner: inviteWallet.ownership.keys.address,
-        });
-      }
-
-      setStoredInvites(ls, invites);
-      console.log('getting invites from code generation');
-      getInvites();
-      showNotification(
-        `Your invite${invites.length > 1 ? 's have' : ' has'} been generated!`
-      );
-    },
-    [
-      api,
-      authToken,
-      connector,
-      contracts,
-      pointCursor,
-      point,
-      wallet,
-      walletType,
-      web3,
-      getDetails,
-      ls,
-      getInvites,
-      setInviteGeneratingNum,
     ]
   );
 
@@ -1049,9 +781,7 @@ export default function useRoller() {
     checkForUpdates,
     config,
     configureNetworkKeys,
-    generateInviteCodes,
     getAndUpdatePoint,
-    getInvites,
     getPoints,
     getPointsDetails,
     getPendingTransactions,
